🚚 extract chat utils
